requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.0
solana==0.30.2
aiohttp==3.9.1 